pyproject.toml
[project]
name = "helloworldcf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.25.3",
    "langchain-azure-ai>=1.2.1",
    "langchain-core>=1.2.27",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
]
